Add @Persist
@Persist is an annotation which annotates fields which should be
persisted.

The single parameter (value, or simply the default) specifies the name
of the tag in which the field is persisted. This does *not* default to
the field's name to encourage forwards/backwards compatibility. It must
be a valid XML tag name and must be unique in its type hierarchy.