Add XmlPersistor, PersistenceUtil
XmlPersistor<R> is a class which is tasked with persisting and
regenerating objects of type R to and from XML. Its constructor takes
a Class<R> and verifies it. It has two methods: toXml(R root) takes an
object of type R and persists it, returning a Document, while
fromXml(Document doc) regenerates an object of type R from a Document.
Neither are implemented as of yet.

PersistenceUtil is a package-level utility class which contains utility
methods for dealing with persistence. Currently, it contains two
methods: verifyAndGetPersistable(Class<?> cls) verifies the existence
and correctness of the @Persistable annotation on cls, then returns it,
while isValidXmlTag(String tag) returns whether its parameter is a valid
XML tag.

As well, update the documentation in README.md and PersistenceException
to reflect that PersistenceException will be the sole custom exception
and that XmlPersistor's constructor has arguments.