Add an element abstraction hierarchy
Add class hierarchy which abstracts the org.w3c.dom library for our
specific purposes.

PersistedElement is the root class. It simply has a tag name and
defines a method toXml(Document) which creates an org.w3c.dom.Element
representing the element.

TextElement is a PersistedElement subclass. It contains text and its
toXml(Document) method reflects that.

ParentElement is a PersistedElement subclass. It has a list of child
PersistedElements whose toXml(Document) methods are called during the
ParentElement's toXml(Document) method to add to the element.

ToplevelElement is a ParentElement subclass. It has two attributes:
root, which is true if this is a root element, and id, which is a String
containing the ID by which the element may be referenced by other
elements. toXml(Document) adds these as attributes, with root being
present only if root="true".

This abstraction hierarchy will be used during XmlPersistor's toXml()
and fromXml() methods.