Add PersistenceStrategy<T>
It's an abstract class whose implementations will provide strategies for
persisting objects of various types. These will include primitives,
strings, and complex objects marked @Persistable. Read the javadoc. This
will be used for dynamically persisting various objects in XmlPersistor.

Note: this does *not* currently provide facilities for accessing the
global toplevel list, so implementations will currently generate
duplicate toplevel elements, which is the opposite of the point. This
will be fixed.