Add PersistablePersistStrategy
It's a PersistenceStrategy which persists each field inside the object
it's passed which is marked @Persist. Currently it cannot persist null
fields and just throws an NPE. It returns either a ParentElement
containing the fields (persisted) or adds them in a ToplevelElement to
the ToplevelList and returns a TextElement comprising a reference to
that ToplevelElement. See the Javadoc of persist() for more details.