Add a test for XmlPersistor.toXml()
The test tests the generation of a simple XML file in which the class
structure contains only primitives and Strings marked @Persist. Several
problems with access control were uncovered and fixed. XMLUnit is used
to actually compare the XML, as to do otherwise is pure torture.

Also, the README was updated to be more consistent with the current API.