Add DuplicatePersistableChecker, PersistingState
The purpose of the class is to detect and reject duplicate tags in
classes marked @Persistable(toplevel=true). This *probably* would not
cause many issues, but throwing PersistenceException makes more sense,
and issues would occur if any IDs were the same among the objects of the
classes with duplicate tags.

The class has a single method, checkAndRegister(), which takes a tag
and a class, checks that the tag does not duplicate anything in its
internal map of previously registered tags and classes (throwing
PersistenceException otherwise), and registers the tag/class
combination.

DuplicatePersistenceChecker instances are global to each call of
toXml(), much like ToplevelList instances. To avoid passing around two
separate instances, PersistingState was created to address this issue.
An instance of PersistingState contains all of the state global to each
toXml() call; this currently includes an instance of ToplevelList and
an instance of DuplicatePersistenceChecker.

A test was also added to check that duplicate @Persistable tags are
rejected.