Support persisting null fields
NullPersistStrategy and NullElement have been added for the purpose of
persisting null fields. A null field is persisted by NullPersistStrategy
like so: <tag null="true" />. This is represented as a NullElement.

Since NullPersistStrategy may replace either PrimitivePersistStrategy
(in the case of Strings) or PersistablePersistStrategy, the latter's
pickStrategy() static method was moved to PersistenceUtil and now
takes the object as an argument as well. NullPersistStrategy is
returned if the object is null. There is also a version which takes
only the class and will never return NullPersistStrategy for use in
e.g. list persisting (when that eventually occurs).

Note that we still do not support persisting null objects directly
since that would be more complicated.