Load XML document into PersistedElement hierarchy
There are now faculties in place in XmlPersistor.fromXml() and the
various PersistedElement subclasses to create a PersistedElement
hierarchy from base XML elements. This is done to ease regeneration.

XmlPersistor.fromXml() loads each ToplevelElement (which in turn loads
all its sub-elements) into a ToplevelList, which will be used in the
future.

Each PersistedElement subclass has a static fromXml(Element) method
which creates a PersistedElement subclass from an XML element.