Add RegenToplevelRegistry
The registry's purpose is to hold the regenerated toplevel objects so
that they can easily be referenced and embedded in each other.

The registry, like the ToplevelList, will be passed around in the
regeneration process, so analogously to PersistingState, RegenState is a
class whose purpose is simply to hold all the state that is global to a
call to XmlPersistor.fromXml().

Also, the registry uses ItemID like ToplevelList, so that class has been
separated into a top-level class.