Add regeneration to PersistenceStrategy
Since regeneration will occur in the same categories as persistence,
PersistenceStrategy has been changed to PersistRegenStrategy and updated
to include a regenerate() method. This method is the reverse of the
persist() method: it takes a PersistedElement and returns the
regenerated object corresponding to it.

All of the former PersistStrategies have been renamed to PRStrategies,
but only PrimitivePRStrategy has its regenerate() method currently
implemented. Note that due to generics, PrimitivePRStrategy.regenerate()
will have to return regenerated primitives as boxed types.

Also, the Objenesis library has been included in build.gradle for
future use in PersistablePRStrategy.