Add simple negative XmlPersistor.fromXml() tests
The tests use a common base, each testing for a specific erroneous
scenario.

Also, make XmlPersistor.fromXml() rethrow any generated
PersistenceExceptions as RegenerationExceptions with the same
message and cause so that PersistenceExceptions are not thrown in the
regeneration method. Common functions between persisting and
regeneration generally throw PersistenceException rather than
RegenerationException, so this is more consistent.