Add tests for objects with no @Persist fields
The XmlPersistor test suite did not cover the case of objects without
fields marked @Persist. Make it so. Fix a bug to do with this in
PersistablePRStrategy.

Also, update README.md to include a (just discovered) note that the
"default values" to which non-@Persist fields are reset is the default
value for the type - 0 or null.